requirements.txt
flask
ultralytics
easyocr
opencv-python-headless
numpy
pillow
gunicorn
